refactor state tables into their own modules. (#329)
