+++ arroyo-worker/src/connectors/polling_http.rs
@@ -17,7 +17,7 @@ use tokio::time::MissedTickBehavior;
 
 use arroyo_rpc::grpc::StopMode;
 use arroyo_rpc::types::Format;
-use arroyo_state::tables::GlobalKeyedState;
+use arroyo_state::tables::global_keyed_map::GlobalKeyedState;
 use tracing::{debug, info, warn};
 use typify::import_types;
