+++ arroyo-worker/src/operators/tumbling_top_n_window.rs
@@ -7,7 +7,7 @@ use std::{
 use crate::engine::{Context, StreamNode};
 use arroyo_macro::process_fn;
 use arroyo_rpc::grpc::{TableDeleteBehavior, TableDescriptor, TableType, TableWriteBehavior};
-use arroyo_state::tables::TimeKeyMap;
+use arroyo_state::tables::time_key_map::TimeKeyMap;
 
 use arroyo_types::*;
 use std::time::Duration;
